transform/requirements.txt
anthropic==0.45.2
fastapi==0.115.8
uvicorn[standard]==0.34.0
pymongo==4.11.1
python-dotenv==1.0.1
pydantic==2.10.6
